+++ tests/dummy/app/controllers/components-acceptance-tests/flexberry-lookup/settings-example-events.js
@@ -1,0 +1,3 @@
+import EditFormController from 'ember-flexberry/controllers/edit-form';
+
+export default EditFormController;