+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+* Event `lookupOnDataLoaded` for `flexberry-lookup` component to detect time of data loading completed.
 
 ## [2.6.1] - 2021-07-02
 ### Fixed